feat: modularize UI into reusable components
